mod: move the zego call init to DI
