+++ app/src/main/java/com/mindgate/mindgateapp/di/SessionModule.kt
@@ -1,11 +1,29 @@
 package com.mindgate.mindgateapp.di
 
+import android.app.Application
+import android.util.Log
+import com.mindgate.mindgateapp.BuildConfig
 import com.mindgate.mindgateapp.data.Session.MongoSessionRepository
 import com.mindgate.mindgateapp.data.repo.SessionsRepository
+import com.zegocloud.uikit.internal.ZegoUIKitLanguage
+import com.zegocloud.uikit.prebuilt.call.ZegoUIKitPrebuiltCallService
+import com.zegocloud.uikit.prebuilt.call.core.invite.ZegoCallInvitationData
+import com.zegocloud.uikit.prebuilt.call.event.CallEndListener
+import com.zegocloud.uikit.prebuilt.call.event.ErrorEventsListener
+import com.zegocloud.uikit.prebuilt.call.event.SignalPluginConnectListener
+import com.zegocloud.uikit.prebuilt.call.event.ZegoCallEndReason
+import com.zegocloud.uikit.prebuilt.call.invite.ZegoUIKitPrebuiltCallInvitationConfig
+import com.zegocloud.uikit.prebuilt.call.invite.ZegoUIKitPrebuiltCallInvitationConfig.generateDefaultConfig
+import com.zegocloud.uikit.prebuilt.call.invite.internal.ZegoTranslationText
+import com.zegocloud.uikit.prebuilt.call.invite.internal.ZegoUIKitPrebuiltCallConfigProvider
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
 import dagger.hilt.components.SingletonComponent
+import im.zego.zim.enums.ZIMConnectionEvent
+import im.zego.zim.enums.ZIMConnectionState
+import org.json.JSONObject
+import timber.log.Timber
 import javax.inject.Singleton
 
 @Module
