+++ app/src/main/java/com/mindgate/mindgateapp/di/ZegoCallManager.kt
@@ -1,0 +1,71 @@
+package com.mindgate.mindgateapp.di
+
+import android.app.Application
+import android.content.Context
+import android.util.Log
+import com.zegocloud.uikit.internal.ZegoUIKitLanguage
+import com.zegocloud.uikit.prebuilt.call.ZegoUIKitPrebuiltCallService
+import com.zegocloud.uikit.prebuilt.call.core.invite.ZegoCallInvitationData
+import com.zegocloud.uikit.prebuilt.call.event.CallEndListener
+import com.zegocloud.uikit.prebuilt.call.event.ErrorEventsListener
+import com.zegocloud.uikit.prebuilt.call.event.SignalPluginConnectListener
+import com.zegocloud.uikit.prebuilt.call.invite.ZegoUIKitPrebuiltCallInvitationConfig
+import com.zegocloud.uikit.prebuilt.call.invite.ZegoUIKitPrebuiltCallInvitationConfig.generateDefaultConfig
+import com.zegocloud.uikit.prebuilt.call.invite.internal.ZegoTranslationText
+import com.zegocloud.uikit.prebuilt.call.invite.internal.ZegoUIKitPrebuiltCallConfigProvider
+import dagger.hilt.android.qualifiers.ApplicationContext
+import im.zego.zim.enums.ZIMConnectionEvent
+import im.zego.zim.enums.ZIMConnectionState
+import org.json.JSONObject
+import timber.log.Timber
+import javax.inject.Inject
+import javax.inject.Singleton
+
+@Singleton
+class ZegoCallManager @Inject constructor(
+    @ApplicationContext private val context: Context
+) {
+
+    fun initZegoInviteService(appID: Long, appSign: String, userID: String, userName: String) {
+        val application = context as Application
+
+        val callInvitationConfig = ZegoUIKitPrebuiltCallInvitationConfig()
+        callInvitationConfig.translationText = ZegoTranslationText(ZegoUIKitLanguage.ENGLISH)
+
+        callInvitationConfig.provider =
+            ZegoUIKitPrebuiltCallConfigProvider { invitationData: ZegoCallInvitationData ->
+                // Make sure generateDefaultConfig is moved inside this class or is accessible
+                generateDefaultConfig(invitationData)
+            }
+
+        ZegoUIKitPrebuiltCallService.events.errorEventsListener =
+            ErrorEventsListener { errorCode: Int, message: String ->
+                Timber.d("onError() called with: errorCode = [$errorCode], message = [$message]")
+            }
+
+        ZegoUIKitPrebuiltCallService.events.invitationEvents.pluginConnectListener =
+            SignalPluginConnectListener { state: ZIMConnectionState, event: ZIMConnectionEvent, extendedData: JSONObject ->
+                Timber.d("onSignalPluginConnectionStateChanged() called with: state = [$state], event = [$event], extendedData = [$extendedData]")
+            }
+
+        ZegoUIKitPrebuiltCallService.init(
+            application, appID, appSign, userID, userName, callInvitationConfig
+        )
+
+        ZegoUIKitPrebuiltCallService.enableFCMPush()
+
+        ZegoUIKitPrebuiltCallService.events.callEvents.callEndListener =
+            CallEndListener { callEndReason, jsonObject ->
+                Log.d(
+                    "CallEndListener",
+                    "Call Ended with reason: $callEndReason and json: $jsonObject"
+                )
+            }
+    }
+
+
+    // Clean up method
+    fun logout() {
+        ZegoUIKitPrebuiltCallService.unInit()
+    }
+}
